Add missing ACL signalling command processing routines.
Add PIN CODE and LINK KEY request processing to the HCI layer - NAK request for the PIN KEY if one is not given in the stack configuration.

Add HCI connection destruction notification to the ACL layer, so that channels associated with the destroyed connection can be closed.

Re-map unknown device error RGB led code to the general error LED code, so that Magenta can be used for packet processing/system busy.